pyproject.toml
[tool.poetry]
name = "fantasia"
version = "1.0.0"
description = "Descripcion de fantasia aquí"
authors = ["frapercan <frapercan1@alum.us.es>", "otro autor"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
docopt = "^0.6.2"
pyyaml = "^6.0.2"


[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
coverage = {extras = ["toml"], version = "^7.2.1"}
flake8 = "^6.0.0"
flake8-bugbear = "^23.2.13"
taskipy = "^1.10.3"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^1.2.0"


[tool.coverage.run]
omit = ["*/site-packages/*"]

[tool.coverage.report]
fail_under = 100

[tool.taskipy.tasks]
html_docs = "make html -C docs"
lint = "poetry run flake8 fantasia"
coverage = "poetry run coverage run -m --source=fantasia pytest tests && poetry run coverage report -m"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

